Refactorization + LIGHT_VALUE
Code related to Torch transformed into TorchManager class.

LIGHT_VALUE constant has been changed to more accurate in my opinion.
